Add service to reset user's password
