+++ app/Http/Controllers/Auth/ResetPasswordController.php
@@ -4,12 +4,9 @@ namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Auth\ResetPasswordRequest;
-use App\Models\User;
-use Illuminate\Auth\Events\PasswordReset;
+use App\Services\ResetPasswordService;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Password;
-use Illuminate\Support\Str;
 use Illuminate\View\View;
 
 class ResetPasswordController extends Controller
@@ -27,23 +24,7 @@ class ResetPasswordController extends Controller
      */
     public function reset(ResetPasswordRequest $request): RedirectResponse
     {
-        $status = Password::reset(
-            $request->only(['email', 'password', 'token']),
-            function (User $user, string $password) use ($request) {
-
-                // Update user's password
-                $user->update([
-                    'password' => Hash::make($password),
-                ]);
-
-                // Remember user if they ticked the remember me checkbox
-                if ($request->validated('remember')) {
-                    $user->setRememberToken(Str::random(60));
-                }
-
-                event(new PasswordReset($user));
-            }
-        );
+        $status = ResetPasswordService::reset($request);
 
         $statusMessage = __($status);
